Refactoring 4
