+++ src/dbnode/storage/index/block.go
@@ -31,6 +31,7 @@ import (
 	"github.com/m3db/m3/src/dbnode/storage/bootstrap/result"
 	"github.com/m3db/m3/src/dbnode/storage/index/compaction"
 	"github.com/m3db/m3/src/dbnode/storage/index/segments"
+	"github.com/m3db/m3/src/dbnode/storage/stats"
 	"github.com/m3db/m3/src/dbnode/tracepoint"
 	"github.com/m3db/m3/src/m3ninx/doc"
 	m3ninxindex "github.com/m3db/m3/src/m3ninx/index"
@@ -970,7 +971,7 @@ func (b *block) addQueryResults(
 	batch []doc.Document,
 ) ([]doc.Document, int, error) {
 	// track recently queried docs to measure memory usage.
-	if err := TrackStats(len(batch)); err != nil {
+	if err := stats.TrackStats(len(batch)); err != nil {
 		return batch, 0, err
 	}
 
@@ -1208,7 +1209,7 @@ func (b *block) addAggregateResults(
 	batch []AggregateResultsEntry,
 ) ([]AggregateResultsEntry, int, error) {
 	// track recently queried docs to measure memory usage.
-	if err := TrackStats(len(batch)); err != nil {
+	if err := stats.TrackStats(len(batch)); err != nil {
 		return batch, 0, err
 	}
